refactor: 모달 기능 관심사 분리
